Add new changes from main

+++ app/src/main/java/com/example/taam_project/RemovePopupFragment.java
@@ -7,16 +7,10 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
-import androidx.fragment.app.DialogFragment;
-
-import android.content.DialogInterface;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
 
-import com.bumptech.glide.Glide;
-import com.google.android.exoplayer2.ExoPlayer;
-import com.google.android.exoplayer2.MediaItem;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -74,6 +68,9 @@ public class RemovePopupFragment extends DialogFragment {
             }
         });
         StorageReference media = sb.child(item.getLotNumber());
+
+        if (item.getMedia().equals("")) return;
+
         media.delete().addOnSuccessListener(unused -> {
             AlertFragment.newInstance("Successfully removed " + item.getName()).show(((MainActivity) context).getSupportFragmentManager(), "alert_fragment");
         }).addOnFailureListener(e -> {
